Show "CLUE" document workspace in Dataflow application Workspace tab [#166986394]
