+++ src/clue/components/clue-app-content.sass
@@ -9,46 +9,3 @@
     right: 0
     bottom: 0
     opacity: 0
-
-  .single-workspace
-    position: absolute
-    top: $header-height + $half-nav-expanded-peek
-    left: $left-nav-contracted-width + $nav-expanded-peek + $workspace-toolbar-width
-    right: $right-nav-contracted-width + $nav-expanded-peek
-    bottom: $bottom-nav-contracted-height
-
-  .left-workspace
-    position: absolute
-    top: $header-height + $half-nav-expanded-peek
-    left: $left-nav-contracted-width + $nav-expanded-peek + $workspace-toolbar-width
-    right: 50%
-    bottom: $bottom-nav-contracted-height
-
-    .workspace
-      margin-right: $half-margin
-
-  .right-workspace
-    position: absolute
-    top: $header-height + $half-nav-expanded-peek
-    left: 50%
-    right: $right-nav-contracted-width + $nav-expanded-peek
-    bottom: $bottom-nav-contracted-height
-
-    .workspace
-      margin-left: $half-margin
-
-    .comparison-placeholder
-      position: absolute
-      top: 0
-      left: 0
-      right: 0
-      bottom: 0
-      color: #000
-      background-color: #F8FCFE
-      margin: $margin
-      border: $border-width solid $border-color
-      border-radius: $border-radius
-      display: flex
-      justify-content: center
-      align-items: center
-      text-align: center
